Added the ElementAndInteractionPrototype scene. Implemented first person control functionality.
ElementAndInteractionPrototype.unity: added to prototype the first person character, use of elements, and general interaction like zooming to a puzzle.

Scripts.meta: since there wasn't anything in the folder, the meta data got updated, hopefully that doesn't mess with other collaborators' versions.

CameraControl: added to let the attached camera follow a given gameobject at a given offset from the given gamobject's frame of reference. The y-axis is also based on the rotation of the given gameobject. The x-axis is based on axis input and a given sensitivity, and is limited by the given upper and lower bounds where 0 is 0 degrees and 1 is 360 degrees.

PlayerControl: added to let the player move the Gameobject in first person. Takes horizontal and vertical (left, right and forward, backward) input from the given axes. Moves at a given speed. Rotates based on a given input axis (horizontal mouse movement) with a given sensitivity.

ProjectVersion.txt: apparently the previous attempt at updating the version only added the Package/Manifest.json file and didn't update this one.