Added a simple interactable lever. Moved camera code into the PlayerControl script.
Lever.prefab: uses the InteractableLever script to act as a lever on a pedestal.

ElementAndInteractionPrototype.unity: added the Lever prefab. Adjusted rigidbody and physics setting to have a slow fall again, but nicer xz movement.

CameraControl: removed because the code was moved into the PlayerControl script.

Interaction folder: added to hold scripts related Interaction functionality.

Interactable: added as the base class for gameobjects that can be interacted with. Allows the raycaster to just call the universal function Interact().

InteractableLever: added to be an interactable with a rotating piece that's rotated when switched on/off.

PlayerControl: moved the CameraControl code into this script for understandability.

TagManager.asset: added the Interaction layer that interactable gameobjects are on.