Added the prototype of the puzzle for the fireplace. Fixed some small parts.
AnimationRelated folder: will holds every Animation, AnimationController, and Behavior.

FireplaceButtonAnimCont: runs a simple animation to show the button being pressed when the parameter isBeingPressed.

PushIn anim: moves the Gameobject backward then forward in the z-axis.

FireplaceButtonResetBeh: added to stop the state transition from occurring again.

FireplaceButtonHolder.prefab: a button in the fireplace puzzle.

LightBase.prefab: a light indicating what a button will do.

PlayerRelatedObjects.prefab: updated with the latest code for elements and element selection.

FirstThirdOfRoom scene: added to start developing the fountain and fireplace puzzles.

InteractionPrototype scene: updated with latest code for elements and element selection.

InteractableFireplaceButton: calls the ReceiveButtonPress function of the FireplacePuzzleController in the given gameobject, passing along the button's number.

InteractableLever: changed the debug key from E to N.

FireplacePuzzleController: holds the logic of the puzzle; buttons correspond to lights, buttons must be pressed in sequence as the lights they relate to are organized in the given array. If the sequence is broken, then the sequence is reset.