Added the LevelSection prefab. Fixed the player horizontal movement. Small name change.
FireplaceButtonHolder prefab: no noticeable changes.

FirstThirdOfRoom scene: added the LevelSection prefab and moved the player character and fire place puzzle over to it.

InteractableFireplaceButton: changed p in fireplace to lower case for consistency.

PlayerControl: transform.TransformPoint appeard to be creating the desired functionality for movment but once player was moved to a different area in the scene  it no longer worked, so now using the correct transformation with transform.TransformDirection.

FireplacePuzzleController: changed p in fireplace to lower case for consistency.

Scenes: no noticeable change.