Added the interaction indicator UI with basic functionality.
InteractionIndicator sprite: simple white ring with black outline. Should consider removing the outline.

PlayerRelatedObjects prefab: added a canvas with an Image of the InteractionIndicator sprite. PlayerController added an InteractionIndicatorUIHolder serialized field that needed to be set to the new canvas image UI element.

FirstThirdOfRoom scene: just the change in the PlayerRelatedObjects prefab being applied to the instance in the scene.

PlayerControl: now boxcasts every frame and enables the InteractionIndicator if an interactable is hit.