Added the progression system and wellspring (for fireplace). Moved Elements enum. Improved visibility of fireplace puzzle lights.
LightBase prefab: changed the material over to water to more easily see the lights.

PlayerRelatedObjects prefab: moved all the canvas elements to under the PlayerRelated gameobject.

FirstThirdOfRoom scene: added the fireplace and progression system.

ElementAffected: since Elements moved to the global scope it's just directly referenced.

Soot: since Elements moved to the global scope it's just directly referenced.

WaterStream: since Elements moved to the global scope it's just directly referenced.

ElementManager: moved Elements to its own script.

Elements: where the enum ended up.

InteractableFireplacButton: moved folders. Now isn't interactable for an amount of time that's assumed to be at least about the same as the animation.

InteractableWellspring: is an Interactable that activates itself when the given progression mark is marked true. Restores the mana of the given element when interacted with.

PlayerControl: since Elements moved to the global scope it's just directly referenced. Used a debug line to make sure interaction indicating was working.

ProgressionMarks: public enum of every progression mark, used to setup and change the progression system's dictionary.

ProgressionSubscribed: interface for classes that want to be notified when a progress mark is changed.

ProgressionSystem: wraps a dictionary of string bool pairs with some extra functionality. Creates the pairs based on the ProgressionMarks enum with bools of false. Dictionary entries can be set to true to mark them as complete. When an entry is marked every subscribed gameobject in the given array has its interface required function called so that they can check things based on the progression system.

FireplacePuzzleController: now marks the fireplace puzzle complete in the progression system when the puzzle is completed.

MeterManager: now has a public method to tell it to restore a passed element type by a given amount.