Added Earth Puzzle Block Asset
I added the "BlockPiece" asset for the earth puzzle. It is EXTREMELY simple for the time being so that the proof of concept can be completed. It can be found in the Assets/Assets/Environment/EarthPuzzle folder
(note: I really should clean up the folder structure at some point but I don't want to accidentally break everything)
Also, I recommend putting a unity material on the block.  It is currently just the default Maya lambert shader.  Maybe go with something brown for a bit of contrast while still meeting the general "earth" feel.