Moved/improved the intro text. Built out the fountain puzzle more. Replaced all Start() with Awake()
Any files without a specific description of their update below have only had their Start functions replaced with Awake.

WaterGaugeAnimCont animation controller: lowered the speed to look less buggy.

PlayerRelatedObjects prefab: moved the IntroText elements into the canvas in the player character to keep with the impromtu standard for the player character and the UI. Adjusted the Intro UI to better scale with resolution. Only the gameobjcts needed to be activated after the intro are activated since every script now uses awake and thus doesn't need to be enabled.

FirstThirdOfRoom scene: decreased the alcove of the fountain puzzle, added in piping to show it running out and into the base of the fountain, and added a spotlight to highlight the water gauge of the puzzle. Other changes based on prefab changes.