Finished basic version of interaction tutorial. Marked progression system users with ProgressionUser interface. Started ElementTutorial scene.
Added a BasicButton animation, animation controller, and behaviour that's similar to the fireplace buttons.

PlayerRelatedObjects prefab: increased the interact distance.

BasicButtonHolder prefab: is very similar to the fireplace buttons. Used in the interaction tutorial for the player to practice interacting.

ElementTutorial scene: started to create this tutorial scene with the same kind of environment.

InteractionTutorial scene: finished a simple version where just any of the buttons on a panel need to be pressed three times to complete the scene.

MovementTutorial scene: changed the transition to the movement tutorial.

FirstThirdOfRoom scene: renamed fireplace puzzle gameobject name.

LockActivationController, WheelController, InteractableLever, EarthButton: based on renaming of ForInteract to ForPauseInteract.

InteractableFireplaceButton, InteractableValveTurnPoint: name change to ForAwake and ForInteract.

Interactable: includes being able to disable the collider for interaction.

InteractableWellspring: now noted as a ProgressionUser. Interact changed to ForInteract.

KeyDownInteractable: based on base class changes.

PauseForInteractable: based on base class changes.

InteractableTutorialButton: basic depressable button that calls the basic button controller on interaction.

PlayerControl: now noted as a ProgressionUser.

ProgressionMarks: added BasicButtons.

ProgressionUser: used to mark classes that use the progression system.

ProgressionUserList text: list of classes that inherit from ProgressionUser.

TutorialCompletionChecker: now noted as a ProgressionUser.

TutorialProgressChecker: now noted as a ProgressionUser.

FireplacePuzzleController: now noted as a ProgressionUser.

FountainPuzzleController: now noted as a ProgressionUser.

BasicButtonsController: is complete when it receives button presses equal to the number of buttons.

ProgressBasedNextSceneCollider: now noted as a ProgressionUser. Cleaned up.

Scenes: added the ElementTutorial.

EditorBuildSettings asset: added the ElementTutorial to the build order.