Lowered fireplace puzzle difficulty by removing 3 lights. Changed the light materials to allow them to be hidden by the soot.
FireplacePuzzle prefab: saved the puzzle since the instance in the scene was being reduced to 5 buttons and lights.

FirstThirdOfRoom scene: adjusted puzzle to 5 lights and changed the materials for lights to indicator ones used in the final puzzle due to the transparent ones being visible behind the soot.

Soot: cleaned up.