Changes to Wind Puzzle / Trying to Fix Final Puzzle Progression Issues
Changes to Wind Puzzle / Trying to Fix Final Puzzle Progression Issues

The final puzzle has issues with the current setup for the powerController getting its PowerOn() activation as a result of checking the progression system.  Commenting out this code and running the old "debug" version that is still in the file allows it to work completely fine.  Both versions use the same PowerOn() method so the activation should be exactly the same.  However, for some reason it doesn't work when called from within the progression checks.

Changes
-Added a different animator to the fans in the final section of the wind puzzle (above and below the wind wellspring).  They now animate correctly.
- Made the same fans controlled by the button on the panel next to them (as was originally intended)
- Also made the wellspring itself controlled by this button to make it so that it can only ever be activated once (trying to fix final puzzle bug)