Added a splash screen.
SplashImage image: used in the SplashScreen scene as the splash screen.

SplashScreen scene: added to be the first scene after unity splash screen.

ElementSelectionTutorial, ElementTutorial, InteractionTutorial, MovementTutorial scene: adjusted exit door serialized fields because of the addition to the ScenesList.

AnyKeySceneLoader: if any key is pressed it loads the given scene.

Scenes: added the SplashScreen scene to the list at the front.

EditorBuildSettings asset: added the SplashScreen scene to the front of the build order.