Added an EndScreen. Moved the objects related to the end screen into place. Replaced the Splash screen single image with UI elements with base Unity fonts.
CreativeCommonImage image: used to credit the music composer in the EndScreen scene.

EndScreen scene: shown after entering the final door.

PlayTest scene: added the ui and collider to transition to the EndScreen.

SplashScreen scene: replaced powerpoint screencapture with UI elements and base Unity fonts.

ActivateOnEntry: when this collides with a gameobject with the given tag, activate a given gameobject.

AnyKeySceneLoader: changed to key down to prevent immediate scene traversal.

Scenes: added the EndScreen to the list.

EditorBuildSettings asset: added EndScreen to the end of the build order.