Fixed duplicate clues, adjusted ring interaction delay. FIxed use of element wheel while hovering over interactable.
EndScreen scene: adjusted UI elements to prevent overlapping.

Playtest scene: changed duplicate clues, and the letter "D" with hint font isn't legible so it was replaced with "d". Also decreased the wait time for the pause interaction of the final door rings.

ElementTutorial, InteractionTutorial, MovementTutorial scene: adjusted UI hierarchy so that the selection wheel covers the interaction UI when it's in use.