Improved end sequence. Made quit button accessible in every part of the game.
EndScreen scene: adjusted anchors to scale UI correctly. Added a quit button.

Playtest scene: added a midgame info screen that appears when the player enters the room behind the final door that informs that they're done for now. It then loads the EndScreen Scene. Added a quit button to the element selection screen.

ElementSelectionTutorial, ElementTutorial, InteractionTutorial, MovementTutorial scene: added a quit button to the element selection screen.

Splash meta: left over file that needed to be removed.

QuitButton: adds a listener to a button on the Gameobject. When invoked, quits the application.